Module Round_robin_with_priority.Count_zeros

Round-robin arbiter implementation which uses masks and a trailing_zeros counter to find the next active input relative to index and then a mux to select the data. The architecture will work, but be horribly inefficient if the number of input sources is not a power of 2.

val combinational : 'a combinational

val sequential : sequential