Module Rac.Mode

type t =

| Fixed

| Integer

In integer mode the coefficients and accumulator are treated as integers, the internal shift registers shift msb first, and the accumulator shifts to the left. This in turn specifies an exact result, so long as the accumulator is large enough to hold it.

In fixed mode the coefficients and accumulator are treated as fixed point values, the shift regiters shift lsb first and the accumulator shifts to the right.

val sexp_of_t : t -> Sexplib0.Sexp.t